fix: Endless mkamigrations rtchat resolved
body: The issue was with the infinite creation of migrations even
when there were no changes. The problem was due to the default
value for the group_name field, which randomly generated a value.
Now, the random hash value is generated when creating the object,
not in the models

footer: null