refactor: Component/router refactor and some issues
body:
1 - Profile navbar btn issue:
Previously, our click event listener was only checking if the clicked
element was an <a> tag. This caused a problem when clicking on images
or other nested elements inside the <a> tag, as the event target was
not detected as an anchor, resulting in the default browser behavior
(full page reload) being executed.

To resolve this, we've updated the click event handler to use
`event.target.closest("a")` to find the closest ancestor <a> tag,
if present. This ensures that the preventDefault() method is
correctly applied, preventing the default navigation.
footer: null

2 - Chat popup btn issue is solved

3 - Component/router refactor is done - PABLO SORRY!!